+++ src/components/map/MapBaseLayers.tsx
@@ -1,7 +1,8 @@
 'use client'
 
-import { LayersControl, TileLayer, useMap } from 'react-leaflet'
+import { LayersControl, useMap } from 'react-leaflet'
 import { MAP_MAX_ZOOM, MAP_MIN_ZOOM, MAP_TILE_BOUNDS, MAP_TILE_URLS } from '@/lib/map/constants'
+import { MapTileLayer } from './MapTileLayer'
 import './MapBaseLayers.css'
 
 export default function MapBaseLayers() {
@@ -10,12 +11,12 @@ export default function MapBaseLayers() {
   return (
     <LayersControl position="topright" collapsed={false}>
       <LayersControl.BaseLayer checked name="Atlas">
-        <TileLayer url={MAP_TILE_URLS.atlas} noWrap bounds={MAP_TILE_BOUNDS} minZoom={MAP_MIN_ZOOM} maxZoom={MAP_MAX_ZOOM} eventHandlers={{
+        <MapTileLayer url={MAP_TILE_URLS.atlas} noWrap bounds={MAP_TILE_BOUNDS} minZoom={MAP_MIN_ZOOM} maxZoom={MAP_MAX_ZOOM} eventHandlers={{
           add: () => { map.getContainer().style.background = '#0FA8D2' },
         }} />
       </LayersControl.BaseLayer>
       <LayersControl.BaseLayer name="위성">
-        <TileLayer url={MAP_TILE_URLS.satellite} noWrap bounds={MAP_TILE_BOUNDS} minZoom={MAP_MIN_ZOOM} maxZoom={MAP_MAX_ZOOM} eventHandlers={{
+        <MapTileLayer url={MAP_TILE_URLS.satellite} noWrap bounds={MAP_TILE_BOUNDS} minZoom={MAP_MIN_ZOOM} maxZoom={MAP_MAX_ZOOM} eventHandlers={{
           add: () => { map.getContainer().style.background = '#153E6A' },
         }} />
       </LayersControl.BaseLayer>